% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqandrev.R
\name{freqandrev}
\alias{freqandrev}
\title{freqandrev}
\usage{
freqandrev(datatable1, datatable2)
}
\arguments{
\item{database1}{a table that has the number caught for each fish species for each location.}

\item{database2}{ia table that has prices for different fish}
}
\value{
1) Most caught fish at each location. 2) Revenue per location ($) 3) total revenue ($)
}
\description{
summaries fish catches for different locations that take as inputs
}
\author{
Matt Gargiulo
}
