% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growthrate.R
\name{growthrate}
\alias{growthrate}
\title{growthrate}
\source{
https://gauchospace.ucsb.edu/courses/pluginfile.php/2004968/mod_assign/introattachment/0/aquaculture_271-216.pdf?forcedownload=1
}
\usage{
growthrate(T, a, b, c, d)
}
\arguments{
\item{T}{Temperature in Celsius}

\item{a}{parameter}

\item{b}{parameter}

\item{c}{parameter}

\item{d}{parameter}
}
\value{
G (Specific Growhth rate (unitless))
}
\description{
Effects of Temperature and Body Weight on Growth Rate estimated by a third order polynomial
}
\author{
Matt Gargiulo
}
